server/reactivestocks/requirements.txt
Django
django-cors
djangorestframework
django-cors-headers
djangorestframework-simplejwt
requests
asgiref
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
